+++ matholymp/test/test_mo_document_generate.py
@@ -48,11 +48,12 @@ class MoDocumentGenerateTestCase(MoScriptTestCase):
     """
 
     def __init__(self, method_name='runTest', script_dir=None, script=None,
-                 top_dir=None, this_dir=None):
+                 top_dir=None, this_dir=None, coverage=False):
         """Initialise a MoDocumentGenerateTestCase."""
         super(MoDocumentGenerateTestCase, self).__init__(method_name,
                                                          script_dir, script,
-                                                         top_dir, this_dir)
+                                                         top_dir, this_dir,
+                                                         coverage)
         if this_dir is not None and not self.check_dir:
             doc_list_file = os.path.join(self.full_dir, 'doc-list')
             doc_list_text = read_text_from_file(doc_list_file)
@@ -67,4 +68,4 @@ class MoDocumentGenerateTestCase(MoScriptTestCase):
 def load_tests(loader, standard_tests, pattern):
     """Return a TestSuite for all the mo-document-generate tests."""
     return load_script_tests('mo-document-generate',
-                             MoDocumentGenerateTestCase)
+                             MoDocumentGenerateTestCase, loader.coverage)