+++ matholymp/test/test_mo_static_upgrade.py
@@ -44,15 +44,17 @@ class MoStaticUpgradeTestCase(MoScriptTestCase):
     """
 
     def __init__(self, method_name='runTest', script_dir=None, script=None,
-                 top_dir=None, this_dir=None):
+                 top_dir=None, this_dir=None, coverage=False):
         """Initialise a MoStaticUpgradeTestCase."""
         super(MoStaticUpgradeTestCase, self).__init__(method_name,
                                                       script_dir, script,
-                                                      top_dir, this_dir)
+                                                      top_dir, this_dir,
+                                                      coverage)
         if this_dir is not None:
             assert self.check_dir
 
 
 def load_tests(loader, standard_tests, pattern):
     """Return a TestSuite for all the mo-static-upgrade tests."""
-    return load_script_tests('mo-static-upgrade', MoStaticUpgradeTestCase)
+    return load_script_tests('mo-static-upgrade', MoStaticUpgradeTestCase,
+                             loader.coverage)